feat: Refactor UserDataBuilder to accept partial props
This commit refactors the `UserDataBuilder` function in the `user-data-builder.ts` file to accept partial properties for the `UserProps` parameter. Previously, the function only accepted complete properties, but now it can handle partial properties as well. This change improves the flexibility and usability of the `UserDataBuilder` function, allowing for more concise and customizable user data generation.