feat: Create users table in the database
This commit adds the SQL migration script to create the "users" table in the database. The table includes columns for "id", "name", "email", "password", and "createdAt". The primary key constraint is set on the "id" column, and a unique index is created on the "email" column. This change enables the storage and retrieval of user information in the database.