feat: Update UserPrismaRepository to insert user entity into the database
This commit updates the UserPrismaRepository class in the user-prisma.repository.ts file to use the PrismaService to insert a user entity into the database. The insert method now uses the create method of the PrismaService to create a new user record with the data from the entity. This change enables the storage of user information in the Prisma database.