refactor: Improve UserEntity unit tests
This commit refactors the UserEntity unit tests in the user.entity.spec.ts file. The changes include improving the readability and organization of the tests, as well as fixing a bug in the email and password field getters. This refactor enhances the reliability and maintainability of the UserEntity unit tests.