feat: Refactor UsersController to use UserPresenter for user output
This commit refactors the UsersController to use the newly added UserPresenter class for presenting user output. The UserPresenter is responsible for transforming the UserOutput object into a format suitable for the API response. By using the UserPresenter, the UsersController improves the consistency and control of the user output presentation, enhancing the reliability and functionality of the application.