feat: Add UserCollectionPresenter to UsersController
This commit adds the UserCollectionPresenter class to the UsersController. The UserCollectionPresenter is responsible for transforming the ListUsersUseCase.Output object into a format suitable for the API response. By including this class, the application gains the ability to easily present a collection of user data, improving the reliability and functionality of the project.