feat: Update test:e2e script to use jest.e2e.config.ts
This commit updates the test:e2e script in the package.json file to use the jest.e2e.config.ts configuration file instead of the jest-e2e.json file. This change allows for better customization and configuration of the end-to-end tests, improving the reliability and functionality of the project.